+++ front/lib/api/llm/clients/google/utils/google_to_events.ts
